refactor: 사용하지 않는 코드 정리
