+++ src/main/java/project/votebackend/dto/follow/FollowUserDto.java
@@ -9,6 +9,5 @@ public class FollowUserDto {
     private Long userId;
     private String name;
     private String profileImage;
-    private String introduction;
     private boolean isFollowing;
 }
