+++ src/main/java/project/votebackend/dto/user/OtherUserPageDto.java
@@ -13,18 +13,13 @@ import java.time.LocalDateTime;
 @Builder
 public class OtherUserPageDto {
 
-    private String username;
     private String name;
     private String profileImage;
-    private String introduction;
     private String address;
-    private String grade;
     private Long avgParticipantCount;
-    private Long point;
     private Long followerCount;
     private Long followingCount;
     private Long postCount;
-    private Long participatedCount;
     private LocalDateTime createdAt;
     private Page<OtherUserVotes> posts;
 }
