+++ src/main/java/project/votebackend/dto/user/UserPageDto.java
@@ -12,16 +12,13 @@ import java.time.LocalDateTime;
 @Builder
 public class UserPageDto {
 
-    private String username;
     private String name;
     private String profileImage;
-    private String introduction;
     private String grade;
     private String address;
     private Long followerCount;
     private Long followingCount;
     private Long avgParticipantCount;
-    private Long point;
     private Long postCount;
     private Long participatedCount;
     private LocalDateTime createdAt;
