+++ src/main/java/project/votebackend/dto/vote/VoteSummaryDto.java
@@ -18,32 +18,8 @@ public class VoteSummaryDto {
     private Long voteId;
     private String title;
     private String thumbnailImageUrl;
-
     private int totalVotes;              // 누적 투표 수
-//    private int todayVotes;              // 오늘 투표 수
-    private int commentCount;            // 댓글 수
-    private int likeCount;               // 좋아요 수
     private LocalDateTime finishTime;
-
-//    private int rankTotal;               // 누적 투표 기준 랭킹
-//    private int rankToday;               // 오늘 득표 기준 랭킹
-//    private int rankComment;             // 댓글 기준 랭킹
-
-//    private int rankChangeTotal;         // 누적 득표 순위 변화
-//    private int rankChangeToday;         // 오늘 득표 순위 변화
-//    private int rankChangeComment;
-
-//    private int ongoingCommentRank;         // 진행중인 투표 댓글 수 랭킹
-//    private int ongoingVoteCountRank;       // 진행중인 투표 투표 수 랭킹
-//
-//    private int ongoingCommentRankChange;   // 진행중인 투표 댓글 수 랭킹 변화
-//    private int ongoingVoteCountRankChange; // 진행중인 투표 투표 수 랭킹 변화
-//
-//    private int endedCommentRank;         // 종료된 투표 댓글 수 랭킹
-//    private int endedVoteCountRank;       // 종료된 투표 투표 수 랭킹
-//
-//    private int endedCommentRankChange;   // 종료된 투표 댓글 수 랭킹 변화
-//    private int endedVoteCountRankChange; // 종료된 투표 투표 수 랭킹 변화
 
     // 투표에 등록된 첫 이미지 썸네일 반환
     private static String extractThumbnail(Vote vote) {
@@ -70,10 +46,6 @@ public class VoteSummaryDto {
                 .finishTime(vote.getFinishTime())
                 .thumbnailImageUrl(VoteSummaryDto.extractThumbnail(vote))
                 .totalVotes(vote.getSelections().size())
-                .commentCount((int) vote.getComments().stream().filter(c -> c.getParent() == null).count())
-                .likeCount((int) vote.getReactions().stream()
-                        .filter(r -> r.getReaction() == ReactionType.LIKE)
-                        .count())
                 .build();
     }
 }
