+++ src/main/java/project/votebackend/service/storage/StorageService.java
@@ -43,28 +43,4 @@ public class StorageService {
                 .map(VoteSummaryDto::from)
                 .toList();
     }
-
-//    //내가 투표한 게시물
-//    public Page<LoadVoteDto> getVotedPosts(Long userId, Pageable pageable) {
-//        Page<Vote> votes = voteRepository.findVotedByUserId(userId, pageable);
-//        List<Long> voteIds = votes.getContent().stream().map(Vote::getVoteId).toList();
-//        Map<String, Object> stats = voteStatisticsUtil.collectVoteStatistics(userId, voteIds);
-//        return voteStatisticsUtil.getLoadVoteDtos(userId, votes, stats, pageable);
-//    }
-//
-//    //북마크한 게시물
-//    public Page<LoadVoteDto> getBookmarkedPosts(Long userId, Pageable pageable) {
-//        Page<Vote> votes = voteRepository.findBookmarkedVotes(userId, pageable);
-//        List<Long> voteIds = votes.getContent().stream().map(Vote::getVoteId).toList();
-//        Map<String, Object> stats = voteStatisticsUtil.collectVoteStatistics(userId, voteIds);
-//        return voteStatisticsUtil.getLoadVoteDtos(userId, votes, stats, pageable);
-//    }
-//
-//    //내가 작성한 게시물
-//    public Page<LoadVoteDto> getCreatedPosts(Long userId, Pageable pageable) {
-//        Page<Vote> votes = voteRepository.findByUser_UserId(userId, pageable);
-//        List<Long> voteIds = votes.getContent().stream().map(Vote::getVoteId).toList();
-//        Map<String, Object> stats = voteStatisticsUtil.collectVoteStatistics(userId, voteIds);
-//        return voteStatisticsUtil.getLoadVoteDtos(userId, votes, stats, pageable);
-//    }
 }
